JS Text Truncation
This script will truncate text. The Simple Method truncates the text to
an exact length, then adds the ellipsis to the end of it. The Advanced
Method truncates the text to the end of the next word after that exact
length, then adds a space and the ellipsis to the end of it.